refactor: move commit related types into file
